+++ CHANGELOG.md
@@ -22,6 +22,8 @@ since 0.4.1, and this project adheres to [Semantic Versioning](https://semver.or
 
 ## Changed
 
+- Upgrade the Python version we ship in Windows / macOS to 3.13
+  ([#1120](https://github.com/freedomofpress/dangerzone/issues/1120))
 - (Docs) Update installation instructions for Fedora. `dnf config-manager` is not a plugin ([#1176](https://github.com/freedomofpress/dangerzone/pull/1176))
 - (Docs) Update installation instructions (and CI checks) for Debian derivatives ([#1141](https://github.com/freedomofpress/dangerzone/pull/1141),
   [#1163](https://github.com/freedomofpress/dangerzone/pull/1163))
@@ -33,6 +35,7 @@ since 0.4.1, and this project adheres to [Semantic Versioning](https://semver.or
 ### Removed
 
 - Platform support: Drop support for Fedora 41 as security support has ended ([#1178](https://github.com/freedomofpress/dangerzone/issues/1178))
+
 ### Development changes
 
 - Vendor GitHub assets using the `mazette` tool (see [#1146](https://github.com/freedomofpress/dangerzone/issues/1146) for the original implementation, renamed later from `assets` to `mazette`)